Login upgraded

+++ Parkly/Parkly/Files/DTLotExpandedTableViewCell.m
@@ -68,7 +68,7 @@
   
 }
 
--(int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+-(long)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
   return self.theSpots.count;
 }
